App has been split into the parts. Validation structure has been reworked. Light changes and reconfigurations in the code.
