+++ siteweather/templates/authentication/login.html
@@ -12,7 +12,7 @@
 {% endblock %}
 
 {% block content %}
-  <form action="{% url 'weather:login' %}" method="post" autocomplete="off">
+  <form action="{% url 'siteweather:auth:login' %}" method="post" autocomplete="off">
     {% csrf_token %}
     <div class="card mb-4">
       <div class="card-header">
