+++ siteweather/templates/authentication/registration.html
@@ -16,7 +16,7 @@
 
 
 {% block content %}
-  <form action="{% url 'weather:registration' %}" method="post" autocomplete="off">
+  <form action="{% url 'siteweather:auth:register' %}" method="post" autocomplete="off">
     {% csrf_token %}
     <div class="card mb-4">
       <div class="card-header">
