+++ siteweather/templates/profile/password_update.html
@@ -53,7 +53,7 @@
         </div>
         <div class="card-footer text-muted">
           <strong>
-            <a href="{% url 'siteweather:profile_update' %}">Back to editing</a>
+            <a href="{% url 'siteweather:profile:profile_update' %}">Back to editing</a>
           </strong>
         </div>
       </div>
