+++ siteweather/templates/profile/profile.html
@@ -47,7 +47,7 @@
             {% endif %}
             City - {{ profile.user_city }} <br>
             {% if profile.photo %}
-              <img src="{{ profile.photo.url }}" width="100%" alt="profile_photo">
+              <img src="{{ profile.photo.url }}" width="100%" alt="profile_photo" class="img-thumbnail">
             {% endif %}
           </h6>
         </div>
@@ -56,7 +56,7 @@
     <div class="card-footer text-muted">
       <strong>
         {% if profile.username == user.username %}
-          <a href="{% url 'siteweather:profile_update' %}">Edit profile</a> <br>
+          <a href="{% url 'siteweather:profile:profile_update' %}">Edit profile</a> <br>
         {% endif %}
       </strong>
     </div>
