+++ siteweather/urls.py
@@ -1,19 +1,20 @@
-from django.urls import path
+from django.urls import path, include
+
 from .views import *
 
 app_name = 'siteweather'
 
-urlpatterns = [
+siteweather_urls = [
     path('', Home.as_view(), name='home'),
-    path('siteweather/<int:pk>/', ViewCity.as_view(), name='view_city'),
-    path('siteweather/<int:pk>/delete_confirmation/', DeleteCityBlock.as_view(), name='delete_confirmation'),
+    path('<int:pk>/', ViewCity.as_view(), name='view_city'),
+    path('<int:pk>/delete_confirmation/', DeleteCityBlock.as_view(), name='delete_confirmation'),
     path('find_by/', FindCity.as_view(), name='find_by'),
-    path('login/', UserLoginFormView.as_view(), name='login'),
-    path('logout/', UserLogoutView.as_view(), name='logout'),
-    path('registration/', RegisterFormView.as_view(), name='registration'),
-    path('profile/<int:pk>/', UserProfile.as_view(), name='profile'),
-    path('profile/update/', UserProfileUpdate.as_view(), name='profile_update'),
-    path('profile/password_update/', UserPasswordUpdate.as_view(), name='password_update'),
-    path('siteweather/site_settings/', PersonalSiteSettings.as_view(), name='site_settings'),
-    path('siteweather/registered_users/', UsersList.as_view(), name='users_list'),
+    path('registered_users/', UsersList.as_view(), name='users_list'),
+    path('site_settings/', PersonalSiteSettings.as_view(), name='site_settings'),
+]
+
+urlpatterns = [
+    path('', include('siteweather.authentication.urls')),
+    path('', include('siteweather.profile.urls')),
+    path('siteweather/', include(siteweather_urls)),
 ]
