+++ templates/base.html
@@ -26,7 +26,6 @@
       {% endblock %}
       {% block pagination %}
       {% endblock %}
-
     </div>
   </div>
 </div>
