+++ templates/inc/_last_registered_users.html
@@ -1,5 +1,5 @@
 {% for user in last_users %}
-  <a href="{% url 'siteweather:profile' user.pk %}/">
+  <a href="{% url 'siteweather:profile:profile' user.pk %}">
     {{ user.first_name }} {{ user.last_name }} <br>
   </a>
 {% endfor %}
