+++ templates/inc/_pagination.html
@@ -4,11 +4,11 @@
       <li class="page-item">
         {% if 'city_name_filter' in request.get_full_path %}
           <a class="page-link"
-             href="/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ page_obj.previous_page_number }}">
+             href="/siteweather/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ page_obj.previous_page_number }}">
             Previous
           </a>
         {% else %}
-          <a class="page-link" href="?page={{ page_obj.previous_page_number }}">
+          <a class="page-link" href="/siteweather/?page={{ page_obj.previous_page_number }}">
             Previous
           </a>
         {% endif %}
@@ -21,11 +21,11 @@
         <li class="page-item active" aria-current="page">
           {% if 'city_name_filter' in request.get_full_path %}
             <a class="page-link"
-               href="/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ p }}">
+               href="/siteweather/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ p }}">
               {{ p }}
             </a>
           {% else %}
-            <a class="page-link" href="?page={{ p }}">
+            <a class="page-link" href="/siteweather/?page={{ p }}">
               {{ p }} <span class="sr-only">(current)</span>
             </a>
           {% endif %}
@@ -34,11 +34,11 @@
         <li class="page-item">
           {% if 'city_name_filter' in request.get_full_path %}
             <a class="page-link"
-               href="/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ p }}">
+               href="/siteweather/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ p }}">
               {{ p }}
             </a>
           {% else %}
-            <a class="page-link" href="?page={{ p }}">
+            <a class="page-link" href="/siteweather/?page={{ p }}">
               {{ p }}
             </a>
           {% endif %}
@@ -49,11 +49,11 @@
       <li class="page-item">
         {% if 'city_name_filter' in request.get_full_path %}
           <a class="page-link"
-             href="/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ page_obj.next_page_number }}">
+             href="/siteweather/?city_name_filter={{ request.GET.dict.city_name_filter }}&date_filter=&page={{ page_obj.next_page_number }}">
             Next
           </a>
         {% else %}
-          <a class="page-link" href="/?page={{ page_obj.next_page_number }}">
+          <a class="page-link" href="/siteweather/?page={{ page_obj.next_page_number }}">
             Next
           </a>
         {% endif %}
