+++ templates/inc/_sidebar.html
@@ -6,14 +6,15 @@
     Find by city
   </a>
   {% if not user.is_authenticated %}
-    <a href="{% url 'siteweather:registration' %}" class="list-group-item list-group-item-action">
+    <a href="{% url 'siteweather:auth:register' %}" class="list-group-item list-group-item-action">
       Register
     </a>
-    <a href="{% url 'siteweather:login' %}" class="list-group-item list-group-item-action">
+    <a href="{% url 'siteweather:auth:login' %}" class="list-group-item list-group-item-action">
       Login
     </a>
   {% else %}
-    <a href="{% url 'siteweather:profile' user.pk %}" class="list-group-item list-group-item-action border-top-0">
+    <a href="{% url 'siteweather:profile:profile' user.pk %}"
+       class="list-group-item list-group-item-action border-top-0">
       Profile
     </a>
     {% if user.role == 'Premium' or user.is_superuser %}
@@ -21,11 +22,14 @@
         List of users
       </a>
     {% endif %}
-    <form action="{% url 'siteweather:logout' %}" method="get" id="css_form_id">
-      <button type="submit" formmethod="get" class="list-group-item list-group-item-action border-top-0">
-        Logout
-      </button>
-    </form>
+    {#    <form action="{% url 'siteweather:auth:logout' %}" method="get" id="css_form_id">#}
+    {#      <button type="submit" formmethod="get" class="list-group-item list-group-item-action border-top-0">#}
+    {#        Logout#}
+    {#      </button>#}
+    {#    </form>#}
+    <a href="{% url 'siteweather:auth:logout' %}" class="list-group-item list-group-item-action border-top-0">
+      Logout
+    </a>
   {% endif %}
   <a href="{% url 'siteweather:site_settings' %}" class="list-group-item list-group-item-action border-top-0">
     Personal website settings
